cleaned up console logs, minor styling
