+++ src/App.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css?family=Lato:900&display=swap');
+
 .App {
   text-align: center;
   background-color: #E9E9E9;
@@ -15,4 +17,49 @@ ul {
 
 .App-intro {
   font-size: large;
+}
+
+.nav-bar {
+  height: 100%;
+  width: 13rem;
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  /* border-style: solid;
+  border-width: 2px;
+  border-color: red; */
+  background: #042A2B;
+  color: #CDEDF6;
+  /* text-align: start */
+}
+
+li {
+  font-family: 'Lato', sans-serif;
+  text-transform: uppercase;
+  letter-spacing: .2rem;
+  margin-bottom: 1rem;
+  text-align: right;
+}
+
+li:last-child {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+}
+
+ul.nav-list {
+  padding: 1.5rem;
+}
+
+main {
+  font-family: 'Lato', sans-serif;
+  margin-left: 13rem;
+  padding: 2rem;
+}
+
+h3 {
+  text-transform: uppercase;
+  width: 100%;
+  border-bottom: 1px solid #042A2B;
 }
